【代码优化】商城: 完善积分商城

+++ yudao-module-mall/yudao-module-promotion-biz/src/main/java/cn/iocoder/yudao/module/promotion/service/point/PointActivityServiceImpl.java
@@ -1,5 +1,7 @@
 package cn.iocoder.yudao.module.promotion.service.point;
 
+import cn.hutool.core.util.ObjectUtil;
+import cn.iocoder.yudao.framework.common.enums.CommonStatusEnum;
 import cn.iocoder.yudao.framework.common.pojo.PageResult;
 import cn.iocoder.yudao.framework.common.util.object.BeanUtils;
 import cn.iocoder.yudao.module.product.api.sku.ProductSkuApi;
@@ -10,6 +12,7 @@ import cn.iocoder.yudao.module.promotion.controller.admin.point.vo.activity.Poin
 import cn.iocoder.yudao.module.promotion.controller.admin.point.vo.activity.PointActivitySaveReqVO;
 import cn.iocoder.yudao.module.promotion.controller.admin.point.vo.product.PointProductSaveReqVO;
 import cn.iocoder.yudao.module.promotion.dal.dataobject.point.PointActivityDO;
+import cn.iocoder.yudao.module.promotion.dal.dataobject.point.PointProductDO;
 import cn.iocoder.yudao.module.promotion.dal.mysql.point.PointActivityMapper;
 import cn.iocoder.yudao.module.promotion.dal.mysql.point.PointProductMapper;
 import jakarta.annotation.Resource;
@@ -19,14 +22,15 @@ import org.springframework.validation.annotation.Validated;
 import java.util.List;
 import java.util.Map;
 
+import static cn.hutool.core.collection.CollUtil.intersectionDistinct;
+import static cn.hutool.core.collection.CollUtil.isNotEmpty;
 import static cn.iocoder.yudao.framework.common.exception.util.ServiceExceptionUtil.exception;
-import static cn.iocoder.yudao.framework.common.util.collection.CollectionUtils.convertMap;
+import static cn.iocoder.yudao.framework.common.util.collection.CollectionUtils.*;
+import static cn.iocoder.yudao.framework.common.util.collection.MapUtils.findAndThen;
 import static cn.iocoder.yudao.module.product.enums.ErrorCodeConstants.SKU_NOT_EXISTS;
 import static cn.iocoder.yudao.module.product.enums.ErrorCodeConstants.SPU_NOT_EXISTS;
-import static cn.iocoder.yudao.module.promotion.enums.ErrorCodeConstants.POINT_ACTIVITY_NOT_EXISTS;
+import static cn.iocoder.yudao.module.promotion.enums.ErrorCodeConstants.*;
 import static java.util.Collections.singletonList;
-
-// TODO @puhui999: 下次提交完善
 
 /**
  * 积分商城活动 Service 实现类
@@ -47,13 +51,27 @@ public class PointActivityServiceImpl implements PointActivityService {
     @Resource
     private ProductSkuApi productSkuApi;
 
+    private static List<PointProductDO> buildPointProductDO(PointActivityDO pointActivity, List<PointProductSaveReqVO> products) {
+        return BeanUtils.toBean(products, PointProductDO.class, product -> {
+            product.setActivityId(pointActivity.getId()).setActivityStatus(pointActivity.getStatus());
+        });
+    }
+
     @Override
     public Long createPointActivity(PointActivitySaveReqVO createReqVO) {
-        // 1. 校验商品是否存在
+        // 1.1 校验商品是否存在
         validateProductExists(createReqVO.getSpuId(), createReqVO.getProducts());
-        // 插入
-        PointActivityDO pointActivity = BeanUtils.toBean(createReqVO, PointActivityDO.class);
+        // 1.2 校验商品是否已经参加别的活动
+        validatePointActivityProductConflicts(null, createReqVO.getProducts());
+
+        // 2.1 插入积分商城活动
+        PointActivityDO pointActivity = BeanUtils.toBean(createReqVO, PointActivityDO.class)
+                .setStatus(CommonStatusEnum.ENABLE.getStatus())
+                .setStock(getSumValue(createReqVO.getProducts(), PointProductSaveReqVO::getStock, Integer::sum));
+        pointActivity.setTotalStock(pointActivity.getStock());
         pointActivityMapper.insert(pointActivity);
+        // 2.2 插入积分商城活动商品
+        pointProductMapper.insertBatch(buildPointProductDO(pointActivity, createReqVO.getProducts()));
         // 返回
         return pointActivity.getId();
     }
@@ -61,27 +79,92 @@ public class PointActivityServiceImpl implements PointActivityService {
     @Override
     public void updatePointActivity(PointActivitySaveReqVO updateReqVO) {
         // 1.1 校验存在
-        validatePointActivityExists(updateReqVO.getId());
+        PointActivityDO activity = validatePointActivityExists(updateReqVO.getId());
+        if (CommonStatusEnum.DISABLE.getStatus().equals(activity.getStatus())) {
+            throw exception(POINT_ACTIVITY_UPDATE_FAIL_STATUS_CLOSED);
+        }
         // 1.2 校验商品是否存在
         validateProductExists(updateReqVO.getSpuId(), updateReqVO.getProducts());
+        // 1.3 校验商品是否已经参加别的活动
+        validatePointActivityProductConflicts(updateReqVO.getId(), updateReqVO.getProducts());
+
+        // 2.1 更新积分商城活动
+        PointActivityDO updateObj = BeanUtils.toBean(updateReqVO, PointActivityDO.class)
+                .setStock(getSumValue(updateReqVO.getProducts(), PointProductSaveReqVO::getStock, Integer::sum));
+        if (updateObj.getStock() > activity.getTotalStock()) { // 如果更新的库存大于原来的库存，则更新总库存
+            updateObj.setTotalStock(updateObj.getStock());
+        }
+        pointActivityMapper.updateById(updateObj);
+        // 2.2 更新商品
+        updateSeckillProduct(updateObj, updateReqVO.getProducts());
+    }
+
+    @Override
+    public void closePointActivity(Long id) {
+        // 校验存在
+        PointActivityDO pointActivity = validatePointActivityExists(id);
+        if (CommonStatusEnum.DISABLE.getStatus().equals(pointActivity.getStatus())) {
+            throw exception(POINT_ACTIVITY_CLOSE_FAIL_STATUS_CLOSED);
+        }
 
         // 更新
-        PointActivityDO updateObj = BeanUtils.toBean(updateReqVO, PointActivityDO.class);
-        pointActivityMapper.updateById(updateObj);
+        pointActivityMapper.updateById(new PointActivityDO().setId(id).setStatus(CommonStatusEnum.DISABLE.getStatus()));
+        // 更新活动商品状态
+        pointProductMapper.updateByActivityId(new PointProductDO().setActivityId(id).setActivityStatus(
+                CommonStatusEnum.DISABLE.getStatus()));
+    }
+
+    /**
+     * 更新秒杀商品
+     *
+     * @param activity 秒杀活动
+     * @param products 该活动的最新商品配置
+     */
+    private void updateSeckillProduct(PointActivityDO activity, List<PointProductSaveReqVO> products) {
+        // 第一步，对比新老数据，获得添加、修改、删除的列表
+        List<PointProductDO> newList = buildPointProductDO(activity, products);
+        List<PointProductDO> oldList = pointProductMapper.selectListByActivityId(activity.getId());
+        List<List<PointProductDO>> diffList = diffList(oldList, newList, (oldVal, newVal) -> {
+            boolean same = ObjectUtil.equal(oldVal.getSkuId(), newVal.getSkuId());
+            if (same) {
+                newVal.setId(oldVal.getId());
+            }
+            return same;
+        });
+
+        // 第二步，批量添加、修改、删除
+        if (isNotEmpty(diffList.get(0))) {
+            pointProductMapper.insertBatch(diffList.get(0));
+        }
+        if (isNotEmpty(diffList.get(1))) {
+            pointProductMapper.updateBatch(diffList.get(1));
+        }
+        if (isNotEmpty(diffList.get(2))) {
+            pointProductMapper.deleteByIds(convertList(diffList.get(2), PointProductDO::getId));
+        }
     }
 
     @Override
     public void deletePointActivity(Long id) {
         // 校验存在
-        validatePointActivityExists(id);
-        // 删除
+        PointActivityDO pointActivity = validatePointActivityExists(id);
+        if (CommonStatusEnum.ENABLE.getStatus().equals(pointActivity.getStatus())) {
+            throw exception(POINT_ACTIVITY_DELETE_FAIL_STATUS_NOT_CLOSED_OR_END);
+        }
+
+        // 删除商城活动
         pointActivityMapper.deleteById(id);
+        // 删除活动商品
+        List<PointProductDO> products = pointProductMapper.selectListByActivityId(id);
+        pointProductMapper.deleteByIds(convertSet(products, PointProductDO::getId));
     }
 
-    private void validatePointActivityExists(Long id) {
-        if (pointActivityMapper.selectById(id) == null) {
+    private PointActivityDO validatePointActivityExists(Long id) {
+        PointActivityDO pointActivityDO = pointActivityMapper.selectById(id);
+        if (pointActivityDO == null) {
             throw exception(POINT_ACTIVITY_NOT_EXISTS);
         }
+        return pointActivityDO;
     }
 
     /**
@@ -104,6 +187,35 @@ public class PointActivityServiceImpl implements PointActivityService {
             if (!skuMap.containsKey(product.getSkuId())) {
                 throw exception(SKU_NOT_EXISTS);
             }
+        });
+    }
+
+    /**
+     * 校验商品是否冲突
+     *
+     * @param id       编号
+     * @param products 商品列表
+     */
+    private void validatePointActivityProductConflicts(Long id, List<PointProductSaveReqVO> products) {
+        // 1.1 查询所有开启的积分商城活动
+        List<PointActivityDO> activityList = pointActivityMapper.selectList(PointActivityDO::getStatus,
+                CommonStatusEnum.ENABLE.getStatus());
+        if (id != null) { // 更新时排除自己
+            activityList.removeIf(item -> ObjectUtil.equal(item.getId(), id));
+        }
+        // 1.2 查询活动下的所有商品
+        List<PointProductDO> productList = pointProductMapper.selectListByActivityId(
+                convertList(activityList, PointActivityDO::getId));
+        Map<Long, List<PointProductDO>> productListMap = convertMultiMap(productList, PointProductDO::getActivityId);
+
+        // 2. 校验商品是否冲突
+        activityList.forEach(item -> {
+            findAndThen(productListMap, item.getId(), discountProducts -> {
+                if (!intersectionDistinct(convertList(discountProducts, PointProductDO::getSpuId),
+                        convertList(products, PointProductSaveReqVO::getSpuId)).isEmpty()) {
+                    throw exception(POINT_ACTIVITY_SPU_CONFLICTS);
+                }
+            });
         });
     }
